Move form components into their own folder, improve styles
