+++ src/views/webview-app/components/connect-form/authentication/ldap.tsx
@@ -6,7 +6,7 @@ import {
   LDAPPasswordChangedAction,
   LDAPUsernameChangedAction
 } from '../../../store/actions';
-import FormInput from '../form-input';
+import FormInput from '../../form/form-input';
 
 type DispatchProps = {
   onLDAPPasswordChanged: (newPassword: string) => void;
