+++ src/views/webview-app/components/connect-form/authentication/scram-sha-256.tsx
@@ -8,7 +8,7 @@ import {
   PasswordChangedAction,
   UsernameChangedAction
 } from '../../../store/actions';
-import FormInput from '../form-input';
+import FormInput from '../../form/form-input';
 
 const styles = require('../../../connect.module.less');
