+++ src/views/webview-app/components/connect-form/authentication/x509.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { connect } from 'react-redux';
 
 import { ActionTypes, X509UsernameChangedAction } from '../../../store/actions';
-import FormInput from '../form-input';
+import FormInput from '../../form/form-input';
 
 type DispatchProps = {
   onX509UsernameChanged: (newUsername: string) => void;
