+++ src/views/webview-app/components/connect-form/connection-form.less
@@ -1,0 +1,22 @@
+.connection-form {
+  margin-top: 0em;
+  padding: 15px 0;
+
+  background-color: var(--vscode-editor-background);
+  color: var(--vscode-editor-foreground);
+  font-size: 14px;
+  line-height: 20px;
+
+  display: inline-block;
+  width: 400px;
+}
+
+.connection-form-fields {
+  display: block;
+}
+
+.connection-form {
+  margin: 10px;
+  width: 320px;
+  text-align: left;
+}
