+++ src/views/webview-app/components/connect-form/replica-set-input.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { connect } from 'react-redux';
 
 import { ActionTypes, ReplicaSetChangedAction } from '../../store/actions';
-import FormInput from './form-input';
+import FormInput from '../form/form-input';
 
 type DispatchProps = {
   onReplicaSetChanged: (newReplicaSetName: string) => void;
