+++ src/views/webview-app/components/overview-page/overview-page.less
@@ -10,8 +10,10 @@
   position: absolute;
   top: 54px;
   min-width: 700px;
+  margin: 0 auto;
   box-shadow: 4px 4px 7px rgba(0, 0, 0, 0.25);
   background-color: var(--vscode-editor-background);
+  z-index: 10;
 }
 
 .overview-page-connect-form-modal-back {
@@ -21,6 +23,7 @@
   left: 0;
   right: 0;
   bottom: 0;
+  z-index: 5;
 }
 
 .overview-header-bar {
