+++ src/views/webview-app/connect.module.less
@@ -1,8 +1,4 @@
-@pw: #21313c;
-@alertRed: #cf4a22;
-@warningText: #ffdd49;
-@green: rgb(19, 170, 82);
-@gray: #5d6c74;
+@import "~./styles/_variables";
 
 .page {
   display: flex;
@@ -15,19 +11,6 @@
 
 .panel {
   border: 1px solid transparent;
-}
-
-form {
-  margin-top: 0em;
-  padding: 15px 0;
-
-  background-color: var(--vscode-editor-background);
-  color: var(--vscode-editor-foreground);
-  font-size: 14px;
-  line-height: 20px;
-
-  display: inline-block;
-  width: 400px;
 }
 
 label {
@@ -129,172 +112,9 @@ label {
       font-weight: bold;
     }
   }
+}
 
-  .connect-form {
-    margin: 10px;
-    width: 320px;
-    text-align: left;
-
-    .fields {
-      display: block;
-    }
-
-    .form-group {
-      display: block;
-
-      .form-item {
-        margin: 15px auto 15px;
-        width: 100%;
-
-        .form-toggle {
-          margin-top: 5px;
-
-          div {
-            border: none;
-            background-color: var(--vscode-searchEditor-textInputBorder,
-                rgba(161, 161, 161, 0.5));
-          }
-        }
-
-        &-has-error {
-          .form-control {
-            border: 1px solid red;
-
-            &:focus {
-              border: 1px solid red;
-            }
-          }
-
-          label {
-            color: red;
-
-            i {
-              margin-right: 5px;
-            }
-          }
-
-          .form-item-file-button {
-            border-color: red;
-          }
-        }
-
-        label {
-          .help-icon {
-            display: inline-block;
-            font-size: inherit;
-            text-rendering: auto;
-            margin: 0 0 0 5px;
-            cursor: pointer;
-            color: rgba(255, 255, 255, 0.5);
-
-            :hover {
-              color: #fbb129;
-            }
-          }
-        }
-
-        &-file-button {
-          width: 100%;
-          background: var(--vscode-input-background, transparent);
-          border: 1px solid var(--vscode-searchEditor-textInputBorder, rgba(161, 161, 161, 0.5));
-          border-radius: 3px;
-          color: var(--vscode-editor-foreground);
-          padding: 8px 4px 9px;
-          margin: 0;
-
-          .file-icon {
-            display: inline-block;
-            font-size: inherit;
-            text-rendering: auto;
-            margin: 0 5px 0 0;
-            color: var(--vscode-editor-foreground);
-          }
-
-          i.help {
-            display: inline-block;
-            font-size: inherit;
-            text-rendering: auto;
-            margin: 0 0 0 5px;
-            cursor: pointer;
-            color: rgba(255, 255, 255, 0.5);
-
-            :hover {
-              color: #fbb129;
-            }
-          }
-        }
-      }
-    }
-
-    .form-group-separator {
-      margin-bottom: 20px;
-    }
-  }
-
-  .buttons {
-    margin: 0;
-    margin-top: 20px;
-  }
-
-  .form-control {
-    display: block;
-    padding: 6px 10px;
-    font-size: 14px;
-    height: 32px;
-    line-height: 32px;
-    color: var(--vscode-input-foreground);
-    background-color: var(--vscode-input-background);
-    background-image: none;
-    border: 1px solid var(--vscode-searchEditor-textInputBorder, transparent);
-    width: 100%;
-    border-radius: 3px;
-
-    -webkit-box-sizing: border-box;
-    -moz-box-sizing: border-box;
-    box-sizing: border-box;
-
-    &:focus {
-      border: 1px solid var(--vscode-focusBorder);
-    }
-
-    ::placeholder {
-      color: var(--vscode-input-placeholderForeground);
-    }
-  }
-
-  .file-input {
-    width: 0.1px;
-    height: 0.1px;
-    opacity: 0;
-    overflow: hidden;
-    position: absolute;
-    z-index: -1;
-  }
-
-  .btn {
-    background-color: @green;
-    color: white;
-    padding: 9px 12px 10px;
-    margin: 6px 0 0 0;
-    display: inline-block;
-    text-decoration: none;
-    cursor: pointer;
-    border: none;
-    border-radius: 3px;
-    transition: 200ms all;
-    font-size: 14px;
-    background-color: rgb(19, 170, 82);
-    background-image: linear-gradient(rgb(19, 170, 82), rgb(24, 150, 76));
-    border-radius: 3px;
-    text-decoration: none;
-    transition: all 120ms ease 0s;
-    overflow: hidden;
-    border: 1px solid rgb(21, 130, 66);
-    box-shadow: rgb(21, 130, 66) 0px -1px 0px inset;
-
-    &:hover {
-      background-color: #109648;
-      background-image: linear-gradient(#129f4c, #148040);
-    }
-  }
+.buttons {
+  margin: 0;
+  margin-top: 20px;
 }
