+++ src/views/webview-app/styles/_variables.less
@@ -1,0 +1,5 @@
+@pw: #21313c;
+@alertRed: #cf4a22;
+@warningText: #ffdd49;
+@green: rgb(19, 170, 82);
+@gray: #5d6c74;
