+++ tsconfig.json
@@ -26,8 +26,6 @@
     "dist",
     "out",
     ".vscode-test",
-    "scripts",
-    "src/views/webview-app",
-    "src/test/suite/views/webview-app"
+    "scripts"
   ]
 }
